\name{artificial}
\alias{artificial}
\alias{y.artificial}
\alias{x.artificial}
\title{Hawkins, Bradu, Kass's Artificial Data}
\usage{
data(artificial)
}
\description{
  This data set was generated by Hawkins, Bradu and Kass in the 1984 for illustrating some of the merits of a robust technique. The data set consists of 75 observations in four dimensions (one response and three explanatory variables). The first 10 observations are bad leverage points, and the next four points are good leverage points (i.e., their \bold{x} are outlying, but the corresponding \emph{y} fit the model quite well).
}

\format{
  \code{artificial} is matrix with 75 cases (rows) and 4 variables
  (columns) where the last column is the dependent variable,
  \code{y.artificial} and \code{x.artificial} are also available.

\source{
  Hawkins, D.M., Bradu, D., and Kass, G.V. (1984)
  Location of several outliers in multiple regression data using elemental sets.
  \emph{Technometrics,}
  \bold{26}, 197--208.
}

\seealso{
  Rousseeuw, P.J., and Leroy, A.M. (1987)
  \emph{Robust regression and outliers detection,}
  Wiley.}

%\examples{}

\keyword{datasets}
