\name{cavendish}
\alias{cavendish}
\title{Cavendish's determinations of the mean density of the earth Data}
\usage{
data(cavendish)
}
\description{
The Cavendish's determinations of the mean density of the earth data (relative to that of water) are 29 measures performed in the 1798 using a torsion balance devised earlier by Michell. After the sixth of these determinations, Cavendish changed his experimental apparatus by replacing a suspension wire by one that was stiffer. To further complicate matters, Cavendish erred in taking the mean of al 29 determinations by treating the value 4.88 as if it were in fact 5.88. 
}
\format{
\code{cavendish} is a vector of 29 observations, the first sixth are made before the apparatus replacement.
}
\source{
 Cavendish, H. (1900)
 Experiments to determine the density of the earth,
 \emph{Philosophical Transactions of the Royal Society of London for the year 1798}
 (Part II) \bold{88}, 469--526,
 Reprinted in \emph{The law of gravitation}
 (A.S. Mackenzie, ed.) American, New York.
}

\seealso{
 Stigler, S.M. (1977)
 Do robust estimators work with \emph{real} data?
 \emph{Annals of Statistics,}
 \bold{5}, 1055--1098,
 (with discussion).}


%\examples{}

\keyword{datasets}
