\name{rocky}
\alias{rocky}
\title{Rockwell hardness, 100 coils produced in sequence at a Chicago
  Steel Mill Data}
\usage{
data(rocky)
}
\description{
  This data set is the Rockwell Hardness (measured on Rockwell "B"
  scale) of a sample of 100 steel coins produced in sequence in a Chicago steel mill.
}

\format{
  \code{rocky} is a time serie with 100 elements.
}

\source{
  K.W. Hipel  and A.I. McLeod (1994), Time series modelling of water
  resources and environmental systems, Elsevier, Amsterdam.
}

\keyword{datasets}
