\name{selection}
\alias{selection}
\alias{xdata}
\alias{ydata}
\title{Selection's Data}
\usage{
data(selection)
}
\description{
  This data set consists of 60 observations for two variables (ydata and xdata). The appropriate regression model for the first fiftyfourth obsertvations should be like
\eqn{y = x^2 + e}
the last sixth comes from a different model.
}

\format{
\code{xdata} is a vector which contains \eqn{x,x^2}, while \emph{ydata} contains the dependent variable's observations. 
}
\source{
  Agostinelli, C. (1999) 
  Robust model selection in regression via weighted likelihood methodology, 
  submitted to \emph{Statistics & Probability Letters}. 
}

%\seealso{}
%\examples{}
 
\keyword{datasets}
