\name{wle.normal.multi.summaries}
\title{Summaries and methods for wle.normal.multi}
\usage{
print.wle.normal.multi(x, digits = max(3, getOption("digits") - 3), ...)
}

\alias{print.wle.normal.multi}

\arguments{
 \item{x}{an object of class \code{wle.normal.multi}.}
}

\description{Until now, only print \code{\link{methods}} is available for class \code{wle.normal.multi}. \code{print.wle.normal.multi} print nicely the output.}

\author{Claudio Agostinelli}

\seealso{\code{\link{wle.normal.multi}} a function for estimating normal multivariate location and scale models.}

%\examples{}

\keyword{models}
\keyword{multivariate}
\keyword{robust}
