\name{wle.normal.summaries}
\title{Summaries and methods for wle.normal}
\usage{
\method{print}{wle.normal}(x, digits = max(3, getOption("digits") - 3), ...)
}

\alias{print.wle.normal}

\arguments{
  \item{x}{an object of class \code{wle.normal}.}
  \item{digits}{number of digits to be used for most numbers.}
  \item{...}{further arguments passed to or from other methods.}
}

\description{Until now, only print \code{\link{methods}} is available for class \code{wle.normal}. \code{print.wle.normal} print nicely the output.}

\author{Claudio Agostinelli}

\seealso{\code{\link{wle.normal}} a function for estimating normal location and scale models.}

%\examples{}

\keyword{models}
\keyword{robust}
