\name{wle.onestep.summaries}
\title{Summaries and methods for wle.onestep}
\usage{
print.wle.onestep(x, digits = max(3, getOption("digits") - 3), ...)
}

\alias{print.wle.onestep}

\arguments{
  \item{x}{an object of class \code{wle.onestep}.}
  \item{digits}{number of digits to be used for most numbers.}
  \item{...}{further arguments passed to or from other methods.}  
}

\description{Until now, only print \code{\link{methods}} is available for class \code{wle.onestep}. \code{print.wle.onestep} print nicely the output.}

\author{Claudio Agostinelli}

\seealso{\code{\link{wle.onestep}} a function for one-step estimation in linear models.}

%\examples{}

\keyword{regression}
\keyword{robust}
